==> R/01-get_buffer_info.Rmd <==
---
title: '01 - getting urbanisation data'
author: "Maxime Dahirel"
date:
output:
  html_document:
    theme: yeti
    toc: TRUE
    toc_float: TRUE
    code_download: TRUE
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r packages}
library(tidyverse) # CRAN v1.3.2
library(sf)        # CRAN v1.0-9
library(terra)     # CRAN v1.7-3
library(units)     # CRAN v0.8-1

library(here)      # CRAN v1.0.1

options(mc.cores = 4)
```

This is where we estimate the average Imperviousness Density (% sealed-up substrate) around all our study sites, to use later in the data analysis.

**IMPORTANT**: To be able to run this script, a copy of the relevant tile of the Imperviousness Density 2006 High Resolution layer **must** be present in `data/GIS_layers` and have the name `IMD_2006_020m_eu_03035_d03_E30N20.tif`. You can download a copy from the EU Copernicus website (https://land.copernicus.eu/pan-european/high-resolution-layers/imperviousness/status-maps/2006 [valid link at the time of commit]). Alternatively, a copy has been made available alongside public releases of this repository; download it and move it to the relevant subfolder.
Note that since the output of the present script has already been saved (`data/processed_data/urban_info_IMD.csv`), whether or not you do this should not affect your ability to run the main scripts.

# Data loading and preparation

First we load all the geopackage data for each site, and the IMD raster layer

```{r get_data}
raw_sites <- read_sf(here("data", "GIS_layers", "rennes2004_sites_centroids.gpkg"),
                     layer = "rennes2004_sites_centroids")

IMD <- rast(here("data", "GIS_layers", "IMD_2006_020m_eu_03035_d03_E30N20.tif"))
names(IMD) <- "IMD"

NAflag(IMD) <- 255 # IMPORTANT: 255 = nodata
```

Before going ahead, we need to check something:

```{r crs_issue1}
head(raw_sites)
IMD
```

The CRS are different, and in addition sites are in Lon/Lat. This will slightly mess up the building of the buffers, and through that the estimates of the mean IMD. Slightly only since `st_buffer()` is able to reproject on the flow, but it's not perfect. First, a demo of the issue:

```{r crs_issue2}
st_buffer(raw_sites, dist = set_units(100, "m")) |> st_area() 
# variation from site to site, should not happen as all buffers identical in size
```

and the solution:

```{r crs_issue3}
raw_sites <- raw_sites |>
  st_transform(crs = 3035) # this is the CRS of the IMD layer, so 2 birds one stone

st_buffer(raw_sites, dist = set_units(100, "m")) |> st_area()
# better, all are identical. Still not exactly = pi * r^2, but that's due to the approx used by st_buffer()
# set nQuadSegs argument there higher than default and you eventually get the exact areas
```

We then create all the buffers we need, one per spatial scale of interest:

```{r make_buffers}
buffer_widths <- c(100, 300, 600, 900, 1200, 1500, 1800)

for (i in 1:length(buffer_widths)) {
  buffers <- st_buffer(raw_sites, dist = set_units(buffer_widths[i], "m"))

  write_sf(buffers,
    here("data", "GIS_layers", 
         paste0("rennes2004_buffers_", buffer_widths[i], "m.gpkg")),
    layer = paste0("rennes2004_buffers_", buffer_widths[i], "m")
  )
}
```

We then use these buffers, one scale at a time, to estimate the mean Imperviousness and store that in a table, that we then save

```{r mean_IMD}
urban_info <- tibble(SITE = raw_sites$site)

for (i in 1:length(buffer_widths)) {
  buffers <- read_sf(here("data", "GIS_layers", 
                          paste0("rennes2004_buffers_", buffer_widths[i], "m.gpkg")),
    layer = paste0("rennes2004_buffers_", buffer_widths[i], "m")
  )

  reproj_vector <- project(vect(buffers), IMD) # should not be needed since they are the same

  urban_info[, paste0("meanIMD_", buffer_widths[i], "m")] <- terra::extract(
    x = IMD,
    y = reproj_vector, fun = "mean", na.rm = TRUE
  )$IMD
}

write_csv(
  urban_info,
  here("data", "processed_data", "urban_info_IMD.csv")
)
```
